Moving things around, adding cdc & pane files
